Python_DBX_SQL_warehouse/pyproject.toml
[project]
name = "python-dbx-sql-warehouse"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "databricks-sql-connector>=4.2.5",
    "pandas>=2.3.3",
]
